requirements.txt
aiogram==2.25.1
python-docx
python-pptx
